Move TestProgram to a separate module
